requirements.txt
# Auto-Stock Dependencies

# 데이터 수집
yfinance>=0.2.18

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 시각화
matplotlib>=3.7.0

# 머신러닝 (클러스터링, 스케일링)
scikit-learn>=1.3.0

# 설정 파일
pyyaml>=6.0

# 진행 표시
tqdm>=4.65.0

# 날짜/시간
python-dateutil>=2.8.0
